pyproject.toml
[project]
name = "self-play"
version = "0.1.0"
description = "Diffusion-guided reinforcement learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.11.11",
    "ty>=0.0.1a6",
    "absl-py==2.2.2",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "etils==1.12.2",
    "farama-notifications==0.0.4",
    "fonttools==4.58.0",
    "fsspec==2025.5.0",
    "glfw==2.9.0",
    "gymnasium==1.1.1",
    "gymnasium-robotics==1.3.1",
    "h5py==3.13.0",
    "imageio==2.37.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "kiwisolver==1.4.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mdurl==0.1.2",
    "minari[hf]==0.5.3",
    "mujoco==3.1.6",
    "numpy==2.2.6",
    "packaging==25.0",
    "pettingzoo==1.25.0",
    "pillow==11.2.1",
    "pygments==2.19.1",
    "pyopengl==3.1.9",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "rich==14.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "typer==0.15.4",
    "typing-extensions==4.13.2",
    "zipp==3.21.0",
    "torch>=2.7.0",
    "pympler>=1.1",
    "tensorboard>=2.19.0",
    "psutil>=7.0.0",
    "torchvision>=0.22.0",
]
